requirements.txt
pytest==7.3.1
pytest-html==3.2.0
pytest-xdist==3.3.1
playwright==1.35.0
python-dotenv==1.0.0
allure-pytest==2.13.2
pytest-allure-adaptor==1.7.10
requests==2.31.0
 